requirements.txt
flask ~= 3.1.0
python-slugify ~= 8.0.4